requirements.txt
numpy
pandas
tensorflow
tqdm
Pillow
scikit-learn
